Release version 1.6
ima-evm-utils originally was licensed as GPLv2 preventing it from
being linked with GPLv3-only applications.  To address this limitation,
the ima-evm-utils package and library licenses are updated to
GPL-2.0-or-later and LGPL-2.0-or-later respectively.

libimaevm defines and uses a couple of global variables making it not
concurrency-safe.  To address this issue, defined and passed local
variables, defined new library functions, and deprecated old library
functions.

See the NEWS file for a short summary of changes and the git history
for details.

Signed-off-by: Mimi Zohar <zohar@linux.ibm.com>