build: Fix build when openssl/engine.h is not available anymore
On Fedora Rawhide openssl/engine.h has disappeared, indicating the
disablement of the engine. Therefore check for openssl/engine.h and
disable building with engine support if the file is not found.

In libimaevm.c #include openssl/engine.h only when CONFIG_IMA_EVM_ENGINE
is set.

In the public header file imaevm.h only #include the file when
OPENSSL_NO_STATIC_ENGINE is not set, which is now set in Fedora Rawhide.

Signed-off-by: Peter Robinson <pbrobinson@gmail.com>
Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>
Reviewed-by: Petr Vorel <pvorel@suse.cz>
Signed-off-by: Mimi Zohar <zohar@linux.ibm.com>