Add copies of the license text
Add copies of the license texts so they're distributed in the tar files
for offline viewing.

The GNU license associated with the the SPDX License Identifier GPL-2.0
and GPL-2.0-or-later is the same.  Restore the COPYING file from
https://www.gnu.org/licenses/old-licenses/gpl-2.0.txt.

Add the GNU license associated with the SPDX License Identifier
LGPL-2.0-or-later as COPYING.LGPL from
https://www.gnu.org/licenses/old-licenses/lgpl-2.0.txt.

Note: While removing the file specific licenses and replacing them with
the SPDX License Identifier, the special exception for linking with the
OpenSSL library (WITH cryptsetup-OpenSSL-exception) was removed.  With
the relicensing of OpenSSL 3 or later to the Apache 2 license, the
special exception is no longer necessary.

Fixes: 6adc50f68202 ("Change license to LGPL-2.0-or-later and GPL-2.0-or-later")
Signed-off-by: Peter Robinson <pbrobinson@gmail.com>
Reviewed-by: Petr Vorel <pvorel@suse.cz>
Signed-off-by: Mimi Zohar <zohar@linux.ibm.com>