Have check_ossl_access only check for valid types
Have check_ossl_access only check for valid types in the type field so
that a check on the access_info passes even if support for the type
(engine or provider) is not currently built-in. The functions actually
using engine or provider following the type field will then report an
error if engine or provider is not built-in.

Fixes: a02fb74 ("Add support for OpenSSL provider to the library and evmctl")
Fixes: 5d89393 ("Implement imaevm_signhash library function and deprecate sign_hash")
Signed-off-by: Stefan Berger <stefanb@linux.ibm.com>