Use upload/retrieve artifact instead of cache for pull requests
If tests are executed in the context of a pull request, Github prevents the
usage of the cache for security reasons. Thus, in this case, fall back to
the slower upload/download artifact, which works also for pull requests.

Also remove ${{ }} around matrix.env.TST_ENV, since the variable is
automatically evaluated in the if condition, and explicitly check if the
value is not an empty string.

Signed-off-by: Roberto Sassu <roberto.sassu@huawei.com>