ci: build and install the ibmtss
Only build and install the ibmtss from the git repo, when configured to
use the ibmtss package and the package or tssstartup does not exist.
With this change, the "boot_aggregate.test" works properly.

Building from a git repo requires installing git and ca-certificates.
Running the "boot_aggregate.test" in a ci environment requires a
software tpm.  Attempt to install the swtpm.

- libtss-dev was removed from Debian.
- tssstartup does not exist on Ubuntu.

Signed-off-by: Mimi Zohar <zohar@linux.ibm.com>